requirements.txt
streamlit==1.42.0
reportlab==4.0.8
PyPDF2==3.0.1
Pillow==10.2.0